Added directory structure - basic / algorithm

+++ Basic/standard_io.cpp
@@ -2,11 +2,11 @@
 
 int main(void)
 {
-    int val1;
+    double val1;
     std::cout<<"Enter the first number: ";
     std::cin>>val1;
 
-    int val2;
+    double val2;
     std::cout<<"Enter the second number: ";
     std::cin>>val2;
